import 1.0.11
2003-12-14  James Antill  <james@and.org>

	* NEWS: Release 1.0.11

2003-12-14  James Antill  <james@and.org>

	* src/vstr_sc.c (vstr_sc_fmt_add_all): Add a "%d" and "%d%d" static
	format checker compat. Mainly because GCC "knows" %p can't take a
	precision ".*".
	(vstr__sc_fmt_add_cb_buf): Fix custom formatter to use the Vstr locale
	NULL string.
	(vstr__sc_fmt_add_cb_ptr): Fix custom formatter to use the Vstr locale
	NULL string.
	* src/vstr_sc.c (vstr__sc_fmt_add_cb_ptr): Fix type passed to
	vstr_sc_fmt_cb_beg() in the callback.

2003-12-14  Martin Pool <mbp@hp.com>

	* src/vstr_sc.c (vstr__sc_fmt_num_ipv6_compact): 64bit clean patch.

2003-12-09  James Antill  <james@and.org>

	* include/vstr-extern.h: Made vstr_export_chr() pure.

2003-11-08  James Antill  <james@and.org>

	* src/vstr_add_fmt.c (vstr__fmt_mv_spec): Fix for ICC warning.

2003-11-06  James Antill  <james@and.org>

	* src/fix.c: Build fixes for --enable-noposix-host

git-archimport-id: james@and.org--2004-code/vstr--main--1--patch-41

+++ tst/add/tst_add_netstr2.c
@@ -35,7 +35,7 @@ int tst(void)
   vstr_add_netstr2_end(s3, pos, s3->len);
 
   vstr_del(s2, 1, s2->len);
-  vstr_add_fmt(s2, s2->len, "%0*zu:,", VSTR_AUTOCONF_ULONG_MAX_LEN, 0);
+  vstr_add_fmt(s2, s2->len, "%0*zu:,", VSTR_AUTOCONF_ULONG_MAX_LEN, (size_t)0);
   TST_B_TST(ret, 3,
             !VSTR_CMP_EQ(s3, 1, s3->len, s2, 1, s2->len));
   
@@ -58,7 +58,7 @@ int tst(void)
   } while (!vstr_add_netstr2_end(s3, pos, s3->len));
   tst_mfail_num(0);
 
-  vstr_add_fmt(s2, s2->len, "%0*zu:a,", VSTR_AUTOCONF_ULONG_MAX_LEN, 1);
+  vstr_add_fmt(s2, s2->len, "%0*zu:a,", VSTR_AUTOCONF_ULONG_MAX_LEN, (size_t)1);
   TST_B_TST(ret, 6, !VSTR_CMP_EQ(s3, 1, s3->len, s2, 1, s2->len));
   
   return (TST_B_RET(ret));
